+++ CHANGELOG.md
@@ -32,6 +32,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   provided directories ([#788](https://github.com/JohnnyMorganz/luau-lsp/issues/788))
 - Fixed production VSCode extension failing to find server binary when debugging another VSCode extension ([#644](https://github.com/JohnnyMorganz/luau-lsp/issues/644))
 - The Studio Plugin settings configuration is now stored in TestService instead of AnalyticsService to allow it to persist across sessions ([#738](https://github.com/JohnnyMorganz/luau-lsp/issues/739))
+- The Studio Plugin now correctly sends updates to the language server when an instance changes name or ancestry hierarchy ([#636](https://github.com/JohnnyMorganz/luau-lsp/issues/636))
+- Fixed Studio Plugin leaking Instance connections after disconnecting from server
 
 ## [1.37.0] - 2024-12-14